chore(openapi): move private helper
