+++ test/tesla/open_api_test.exs
@@ -1,10 +1,10 @@
-defmodule Tesla.EnvTest do
+defmodule Tesla.OpenAPITest do
   use ExUnit.Case, async: true
 
-  alias Tesla.Env
+  alias Tesla.OpenAPI
 
   test "merges private data maps from left to right" do
-    assert Env.merge_private([
+    assert OpenAPI.merge_private([
              %{tesla_path_template: :template},
              %{tesla_path_params: :path_params},
              %{tesla_query_params: :query_params, tesla_path_params: :override}
